beta test of answer view

+++ quiz_person.php
@@ -6,13 +6,15 @@ $conn=dbConnect('write');
 if ($_SESSION['role'] != "admin") {
   header("Location: index");
 }
-if (isset($_POST['resetviews']) && $_SESSION['role'] == "admin") {
+if (isset($_POST['resetviews']) && isset($_POST['delview']) && $_SESSION['role'] == "admin") {
   foreach($_POST['delview'] as $person) {
     $sql2 = "DELETE FROM quiz_views WHERE id_fk = " . $person;
     $conn->query($sql2);
   }
+}
+if (isset($_POST['resetviews']) && isset($_POST['markpass']) && $_SESSION['role'] == "admin") {
   foreach($_POST['markpass'] as $person) {
-    $sql2 = "UPDATE users SET quiz_score = 101 WHERE user_id = " . $person;
+    $sql2 = "UPDATE users SET quizscore = 101 WHERE user_id = " . $person;
     $conn->query($sql2);
   }
 }
@@ -24,6 +26,29 @@ $result = $conn->query($sql);
 <title>Quiz Views</title>
 </head>
 <body>
+<?php if (isset($_GET['view'])) {
+?>
+<h1>View Answers for User</h1>
+<p>The correct answer field is only valid as long as the answers have not been changed from the quiz page.</p>
+<table border="2">
+<tr><th>Question #</th>
+<th>Answer</th>
+<th>Correct</th>
+<?php
+  $sql = "SELECT * FROM quiz_answers WHERE user_id_fk = " . $conn->real_escape_string($_GET['view']);
+  $persons_answers = $conn->query($sql);
+  $sql = "SELECT * FROM quiz_questions";
+  $questions = $conn->query($sql);
+  while ($answer = $persons_answers->fetch_assoc() && $question = $questions->fetch_assoc()) {
+?>
+<tr>
+<td><?php echo $question['qnum']; ?></td>
+<td><?php echo $answer['answer']; ?></td>
+<td><?php echo $question['answer']; ?></td>
+</tr>
+<?php } ?>
+<a href="quiz.php">Back to Quiz Page</a>
+<?php } else {?>
 <h1>Quiz Management</h1>
   <form method="post" action="" name="meh">
 <table border="2">
@@ -46,6 +71,7 @@ $result = $conn->query($sql);
 </table>
 <input type="submit" name="resetviews" value="Submit">
 </form>
+<?php } ?>
 <a href="index.php">Home</a>
 </body>
 </html>
